requirements.txt
python-telegram-bot
gigachat
python-dotenv
